Zoom centered on the mouse position
Zooming via the scroll wheel now keeps the position under the curser in
place.

+++ src/T2DMap.h
@@ -130,8 +130,8 @@ public:
     double eSize;
     int mRoomID;
     int mAreaID;
-    int mOx;
-    int mOy;
+    qreal mOx;
+    qreal mOy;
     int mOz;
     bool mShiftMode;
     bool mShowInfo;
